pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "integrityx"
version = "1.0.0"
description = "Financial Document Integrity System with Quantum-Safe Cryptography"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "httpx>=0.28.0",
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "walacor-python-sdk>=0.1.5",
    "pathlib2>=2.3.7",
    "typing-extensions>=4.0.0",
    "jsonschema>=4.0.0",
    "canonicaljson>=1.6.0",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "aiohttp>=3.9.0",
    "pyjwt>=2.9.0",
    "cryptography>=46.0.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.0.0"
]

[tool.setuptools]
packages = ["backend.src"]

[tool.setuptools.package-dir]
"" = "backend"
